requirements.txt
absl-py==1.4.0
addict==2.4.0
aiohttp==3.8.5
aiosignal==1.3.1
ansi2html==1.8.0
appdirs==1.4.4
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
av==10.0.0
backcall==0.2.0
backports.cached-property==1.0.2
beautifulsoup4==4.12.3
bleach==6.1.0
blessed==1.20.0
blessings==1.7
brotlipy==0.7.0
build==0.10.0
cachetools==5.3.1
certifi @ file:///croot/certifi_1690232220950/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
chardet==5.2.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.6
codecov==2.1.13
colorama==0.4.6
coloredlogs==15.0.1
colour-runner==0.1.1
comm==0.1.3
ConfigArgParse==1.7
contourpy==1.0.7
coverage==7.2.7
cryptography @ file:///croot/cryptography_1673298753778/work
cycler==0.11.0
Cython==0.29.36
dash==2.11.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
deepdiff==6.3.1
defusedxml==0.7.1
diff-gaussian-rasterization @ file:///home/pengr/Documents/gaussian-splatting/submodules/diff-gaussian-rasterization
distlib==0.3.7
docker-pycreds==0.4.0
docopt==0.6.2
einops==0.6.1
embreex==2.17.7.post1
executing==1.2.0
fastjsonschema==2.18.0
filelock==3.12.2
fire==0.5.0
Flask==2.2.5
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
fonttools==4.39.0
frozenlist==1.4.0
fsspec==2023.6.0
fvcore @ file:///home/conda/feedstock_root/build_artifacts/fvcore_1671623667463/work
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.16.0
google-auth-oauthlib==0.4.6
gpustat==1.0.0
GPUtil==1.4.0
grpcio==1.51.1
huggingface-hub==0.20.3
humanfriendly==10.0
icecream==2.1.3
idna @ file:///croot/idna_1666125576474/work
imageio==2.31.1
imageio-ffmpeg==0.4.8
importlib-metadata==6.8.0
iopath @ file:///home/conda/feedstock_root/build_artifacts/iopath_1636568816351/work
ipython==8.12.3
ipywidgets==8.1.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter_client==8.6.2
jupyter_core==5.3.1
jupyterlab-widgets==3.0.8
jupyterlab_pygments==0.3.0
kiwisolver==1.4.4
lazy_loader==0.3
lightning-lite==1.8.0
lightning-utilities==0.3.0
llvmlite==0.40.1
lmdb==1.4.1
lpips==0.1.4
lz4==4.3.2
Mako @ file:///croot/mako_1665472421453/work
Markdown==3.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mkl-fft==1.3.1
mkl-random @ file:///home/builder/ci_310/mkl_random_1641843545607/work
mkl-service==2.4.0
mmcv==2.0.1
mmengine==0.8.2
model-index==0.1.11
multidict==6.0.4
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.7.0
nest-asyncio==1.5.7
networkx==3.0
ninja==1.11.1.1
numba==0.57.1
numpy @ file:///croot/numpy_and_numpy_base_1672336185480/work
nvidia-ml-py==11.495.46
oauthlib==3.2.2
open3d==0.17.0
opencv-python==4.7.0.68
opendatalab==0.0.9
openmim==0.3.9
ordered-set==4.1.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pipreqs==0.5.0
piq==0.8.0
platformdirs==3.9.1
plotly==5.14.1
pluggy==1.2.0
plyfile==0.9
pooch==1.7.0
portalocker @ file:///croot/portalocker_1682958299664/work
prefetch-generator==1.0.3
progressbar2==4.2.0
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodome==3.18.0
pyembree==0.1.12
Pygments==2.14.0
pyhocon==0.3.57
PyMCubes==0.1.4
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
pyproject-api==1.5.3
pyproject_hooks==1.0.0
pyquaternion==0.9.9
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil==2.8.2
python-utils==3.5.2
pytools @ file:///home/conda/feedstock_root/build_artifacts/pytools_1690307577186/work
pytorch-lightning==1.8.0
pytorch-msssim==1.0.0
pytorch3d==0.7.5
pytz==2023.3
PyWavelets==1.4.1
PyYAML @ file:///croot/pyyaml_1698096049011/work
pyzmq==26.0.3
referencing==0.30.0
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
requests-oauthlib==1.3.1
retrying==1.3.4
rich==13.4.2
rootpath==0.1.1
rpds-py==0.9.2
rsa==4.9
Rtree==1.0.1
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.10.0
sentry-sdk==1.29.0
setproctitle==1.3.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.0
soupsieve==2.5
stack-data==0.6.2
tabulate @ file:///croot/tabulate_1701354810597/work
tenacity==8.2.2
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
termcolor==2.3.0
threadpoolctl==3.2.0
tifffile==2023.2.3
tinycss2==1.3.0
tomli==2.0.1
torch==1.13.1
torch-efficient-distloss==0.1.3
torch-scatter==2.1.1
torchaudio==0.13.1
torchmetrics==0.11.4
torchsparse @ file:///home/pengr/Documents/torchsparse-2.1.0%2Btorch113cu117-cp310-cp310-linux_x86_64.whl
torchvision==0.14.1
tornado==6.4
tox==4.7.0
tqdm @ file:///croot/tqdm_1679561862951/work
traitlets==5.9.0
trimesh==3.22.5
typing_extensions @ file:///croot/typing_extensions_1669924550328/work
tzdata==2023.3
urllib3 @ file:///croot/urllib3_1673575502006/work
virtualenv==20.24.3
wandb==0.15.7
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.2.3
widgetsnbextension==4.0.8
yacs==0.1.8
yapf==0.40.1
yarg==0.1.9
yarl==1.9.2
zipp==3.16.2
